Rename path optimizer module; minor cleanup
